% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsb_sheets.R
\name{xlsb_sheets}
\alias{xlsb_sheets}
\title{List all sheets in a .xlsb workbook}
\usage{
xlsb_sheets(path)
}
\arguments{
\item{path}{Path to the xlsb workbook}
}
\description{
Analogous to the \code{excel_sheets} function from \code{readxl}. Efficiently gets the sheet names without needing to read any surplus data.
}
\examples{
xlsb_sheets(path = system.file("extdata", "TestBook.xlsb", package = "readxlsb"))

}
